backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
xlrd==2.0.2

numpy
xgboost
lightgbm
category-encoders
langgraph
langchain
python-multipart
joblib